refactor categorization components
